feat(metadata): cookies banner restyled
